Updated comments and abstraction
Updated the comments to make the code a little more clear (it was old
commented stuff that I put there that we no longer need). I also handled
the weird abstraction the maps package has by not including Hawaii and
Alaska within states, but rather as accessed through the world layer in
its database. That's 90's original codebase for ya.